+++ Changelog.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Add mutation from `Date.parse` to more strict parsing methods #448
+* Add mutation from `foo.to_i` to `Integer(foo)` #455
 
 # v0.8.5 2015-09-11